refactor: schema 폴더 위치 수정
