+++ src/schemas/scheduleCreationSchema.ts
